backend/pyproject.toml
[build-system]
requires = ["poetry==1.0.5"]
build-backend = "poetry.masonry.api"


[tool.nitpick]
style = "https://raw.githubusercontent.com/wemake-services/wemake-python-styleguide/master/styles/nitpick-style-wemake.toml"


[tool.poetry]
name = "skyexpo_matchmaking"
description = "The matchmaking app for skyexpo 2020."
version = "0.1.0"
readme = "README.md"
authors = ["Marcin Josiński"]


[tool.poetry.dependencies]
python = "3.8.2"
# TODO: remove after https://github.com/python-poetry/poetry/issues/2042
importlib-metadata = "<1.2.0,>=1.1.3"
django = "^3.0"
djangorestframework = "^3.11"
djangorestframework-camel-case = "^1.1"
djangorestframework-simplejwt = "^4.4"
django-split-settings = "^1.0"
django-axes = "^5.3"
django-csp = "^3.6"
django-health-check = "^3.12"
django-http-referrer-policy = "^1.0"
django-feature-policy = "^3.2"
django-countries = "^6.1"
django-phonenumber-field = "^4.0"
django-cors-headers = "^3.2"
phonenumbers = "^8.12"
glom = "^19.10"
Pillow = "^7.0"
psycopg2 = "^2.8"
gunicorn = "^20.0"
python-decouple = "^3.3"
dump-env = "^1.1"
bcrypt = "^3.1"
sentry-sdk = "^0.14"
pyrsistent = "0.14.2"
celery = "^4.4"
redis = "^3.3"
flower = "^0.9"
drf-yasg = "^1.17"

[tool.poetry.dev-dependencies]
django-debug-toolbar = "^2.1"
django-querycount = "^0.7"
django-extensions = "^2.2"
nplusone = "^1.0"

wemake-python-styleguide = "^0.14"
flake8-pytest = "^1.3"
flake8-pytest-style = "^0.1.3"
flake8-django = "^0.0.4"
nitpick = "^0.22"

pytest = "^5.3"
pytest-django = "^3.8"
pytest-cov = "^2.8"
pytest-randomly = "^3.2"
pytest-deadfixtures = "^2.2"
pytest-testmon = "^1.0"
pytest-timeout = "^1.3"
factory_boy = "^2.12"

mypy = "^0.770"
django-stubs = "^1.5"
djangorestframework-stubs = "^1.1"

ipython = "^7.11"
jupyter = "^1.0"
PrettyTable = "&0.7"

yamllint = "^1.20"
safety = "^1.8"
dotenv-linter = "^0.1"
polint = "^0.4"
dennis = "^0.9"

pre-commit = "^1.21"
